requirements.txt
-e .

aiomcache==0.8.2
aiohttp==3.12.12
marshmallow==3.26.1
msgpack==1.1.1
pytest==8.4.0
pytest-asyncio==0.26.0
pytest-cov==6.2.1
pytest-mock==3.14.1
valkey-glide==1.3.5
